move ModalSavedList to own component
