+++ src/components/ModalSavedItem/styles.scss
@@ -1,0 +1,1 @@
+@import "../../styles/components/modal-saved-item";
