+++ src/components/SavedItem/index.js
@@ -51,23 +51,6 @@ SavedItem.propTypes = {
   uri: PropTypes.string
 }
 
-const ModalSavedItem = ({ handleChange, title, uri }) => (
-  // TODO: styling for checkbox
-  <li className="modal-saved-item">
-    <CheckBoxInput
-      className="checkbox--orange"
-      id={uri}
-      checked={true}
-      label={title}
-      handleChange={handleChange} />
-  </li>
-)
-
-ModalSavedItem.propTypes = {
-  handleChange: PropTypes.func,
-  title: PropTypes.string.isRequired
-}
-
 const SavedItemGroup = ({ handleChange, items, toggleInList, title, uri }) => {
   const listItems = items.map((item, index) =>
     <SavedItem
@@ -95,29 +78,6 @@ SavedItemGroup.propTypes = {
   uri: PropTypes.string.isRequired,
 }
 
-const ModalSavedItemGroup = ({ handleChange, items, title }) => {
-  const listItems = items.map((item, index) =>
-    <ModalSavedItem
-      handleChange={handleChange}
-      key={index}
-      {...item} />
-  );
-  return (
-    <div className="modal-saved-items__item-group">
-      <h3 className="modal-item-group__title">{title}</h3>
-      <ul className="modal-item-group__items">
-        {listItems}
-      </ul>
-    </div>
-  )
-}
-
-ModalSavedItemGroup.propTypes = {
-  handleChange: PropTypes.func,
-  items: PropTypes.array.isRequired,
-  title: PropTypes.string.isRequired,
-}
-
 export const SavedItemList = ({ handleChange, isLoading, items, toggleInList }) => {
 
   const groupItems = items => {
@@ -142,26 +102,4 @@ SavedItemList.propTypes = {
   isLoading: PropTypes.bool.isRequired,
   items: PropTypes.array.isRequired,
   toggleInList: PropTypes.func.isRequired,
-}
-
-export const ModalSavedItemList = ({ handleChange, items }) => {
-  const groupItems = items => {
-    return items.length ? (items.map((item) =>
-      <ModalSavedItemGroup
-        key={item.title}
-        {...item}
-        groupUri={item.uri}
-        handleChange={handleChange} />
-    )) : (<p className="saved-items__empty">No saved items.</p>)
-  }
-  return (
-    <div className="modal-saved-items">
-      {groupItems(items)}
-    </div>
-  )
-}
-
-ModalSavedItemList.propTypes = {
-  handleChange: PropTypes.func,
-  items: PropTypes.array.isRequired
 }
