+++ src/styles/components/_modal-saved-item.scss
@@ -1,0 +1,108 @@
+@import "../styles";
+
+.saved-items__item-group,
+.saved-items__loading,
+.saved-items__empty,
+.modal-saved-items__item-group {
+  @include grid-column(12);
+  border-top: 1px dotted $faded-navy;
+  font-family: $sans-serif-stack;
+}
+
+.saved-items__empty {
+  font-family: $sans-serif-stack;
+  padding: 20px 0px;
+}
+
+.modal-item-group__items {
+  list-style: none;
+  padding-left: 0;
+  font-size: 15px;
+  font-weight: 700;
+  color: $dark-gray;
+}
+
+.item-group__title {
+  font-size: 20px;
+  color: $dark-gray;
+  border-bottom: none;
+  margin-top: 24px;
+  margin-bottom: 30px;
+  padding: 0;
+}
+
+.modal-item-group__title {
+  @extend .item-group__title;
+  font-size: 15px;
+  margin: 20px 0px;
+}
+
+.saved-item {
+  @include grid-column(12);
+  margin-bottom: 40px;
+  float: left;
+  position: relative;
+}
+
+.modal-saved-item {
+  margin-bottom: 15px;
+}
+
+.saved-item__inputs, .modal-saved-item__inputs {
+  @include grid-column(1);
+}
+
+.saved-item__item-description, .modal-saved-item__item-description  {
+  @include grid-column(11);
+  @media screen and (min-width: $break-lg) {
+    @include grid-column(9);
+  }
+}
+
+.saved-item__buttons {
+  @media screen and (min-width: $break-lg) {
+    position: absolute;
+    top: 0;
+    right: 0;
+  }
+}
+
+.saved-item__title {
+  margin: 0 0 5px 0;
+  font-size: 20px;
+}
+
+.modal-saved-item__title {
+  @extend .saved-item__title;
+  font-size: 14px;
+}
+
+@mixin saved-item-text {
+  font-family: $sans-serif-stack;
+  font-size: 14px;
+  color: $dark-gray;
+  line-height: 16px;
+}
+
+.saved-item__date {
+  @include saved-item-text;
+}
+
+.saved-item__description {
+  @include saved-item-text;
+}
+
+.saved-item__found-in {
+  @include saved-item-text;
+  & a {
+    @include saved-item-text;
+    color: $dark-gray;
+    font-weight: normal;
+    text-decoration: underline;
+  }
+}
+
+.saved-item__last-requested {
+  @include saved-item-text;
+  font-style: italic;
+}
